extraindo composable para busca de personagens
