+++ src/components/interfaces.ts
@@ -8,31 +8,8 @@ export interface Meta {
   totalCount: number;
 }
 
-export interface Character {
-  id: string;
-  name: string;
-  image: string;
-}
-
 export interface CallbackFunction {
   (stop?: boolean): void;
-}
-
-export interface CharactersQueryInfo {
-  next: number;
-}
-
-export interface CharactersQuery {
-  info: CharactersQueryInfo;
-  results: Array<Character>
-}
-
-export interface CharactersQueryResult {
-  characters: CharactersQuery;
-}
-
-export interface FilterCharacter {
-  name: string;
 }
 
 export interface InfiniteScrollOptions {
