+++ src/pages/Characters.vue
@@ -18,15 +18,14 @@
 import {
   defineComponent, ref, computed, watch,
 } from 'vue';
-import { useQuery } from '@vue/apollo-composable';
-import gql from 'graphql-tag';
 import SearchBar from '../components/SearchBar.vue';
 import Loader from '../components/Loader.vue';
 import CharacterPreview from '../components/CharacterPreview.vue';
-
+import { useFetchCharacters } from '../composables/fetchCharacters';
 import {
-  CallbackFunction, CharactersQueryResult, FilterCharacter, InfiniteScrollOptions,
+  CallbackFunction, InfiniteScrollOptions,
 } from '../components/interfaces';
+import { FilterCharacter } from '../composables/interfaces';
 
 export default defineComponent({
   name: 'Characters',
@@ -36,94 +35,42 @@ export default defineComponent({
     CharacterPreview,
   },
   setup() {
-    const CHARACTERS_QUERY = gql`
-      query characters($page: Int, $filter: FilterCharacter) {
-        characters(page: $page, filter: $filter) {
-          info {
-            next
-          },
-          results {
-            id,
-            name,
-            image
-          }
-        }
-    }`;
-
     const search = ref('');
-    const page = ref(1);
+    const filter = computed<FilterCharacter>(() => ({ name: search.value }));
+    const {
+      characters,
+      loadMore,
+      loading,
+      next,
+    } = useFetchCharacters(filter);
+
     const infiniteScroll = ref<InfiniteScrollOptions | null>(null);
 
-    const filter = computed<FilterCharacter>(() => ({ name: search.value }));
-
-    const {
-      result, fetchMore, refetch, loading,
-    } = useQuery<CharactersQueryResult>(
-      CHARACTERS_QUERY, () => ({
-        page: page.value,
-        filter: filter.value,
-      }),
-    );
-
-    watch(filter, (newValue) => {
+    watch(filter, () => {
       window.scroll({
         top: 0,
         left: 0,
         behavior: 'smooth',
       });
       infiniteScroll.value?.reset();
-      void refetch({
-        page: 1,
-        filter: newValue,
-      });
     });
 
-    const characters = computed(() => result.value?.characters.results ?? []);
+    const onLoad = async (index: number, done: CallbackFunction) => {
+      if (loading.value) {
+        done();
+        return;
+      }
+      if (next.value) {
+        await loadMore();
+      }
+      done();
+    };
 
-    async function loadMore() {
-      await fetchMore({
-        variables: {
-          page: result.value?.characters.info.next,
-          filter: filter.value,
-        },
-        updateQuery: (previousResult: CharactersQueryResult, { fetchMoreResult }) => {
-          if (!fetchMoreResult) {
-            return previousResult;
-          }
-
-          const previousCharacters = previousResult.characters.results;
-          const newCharacters = fetchMoreResult.characters.results;
-          const updateCharacters = previousCharacters.concat(newCharacters);
-          return {
-            ...previousResult,
-            characters: {
-              ...previousResult.characters,
-              info: {
-                ...fetchMoreResult.characters.info,
-              },
-              results: updateCharacters,
-            },
-          };
-        },
-      });
-    }
-    // tentar usar reeef para restar o scroll
     return {
       search,
-      result,
       infiniteScroll,
-      loadMore,
       characters,
-      async onLoad(index: number, done: CallbackFunction) {
-        if (loading.value) {
-          done();
-          return;
-        }
-        if (result.value?.characters.info.next) {
-          await loadMore();
-        }
-        done();
-      },
+      onLoad,
     };
   },
 });
